datamonkey/requirements.txt
# Core dependencies for datamonkey
fastapi>=0.104.0
httpx>=0.25.0
pydantic>=2.5.0
python-dotenv>=1.0.0
rich>=13.7.0
structlog>=23.2.0
pyyaml>=6.0

# Async support
asyncio-mqtt>=0.16.0
openai>=1.40.0
